refactor: controller + service
